orxa-smartdocs-ai/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
python-dotenv==1.0.1
pymupdf==1.24.5
reportlab==4.2.2
python-docx==1.1.2
Pillow==10.4.0
anthropic==0.34.2
gunicorn==22.0.0
